+++ src/textual/app.py
@@ -2840,20 +2840,23 @@ class App(Generic[ReturnType], DOMNode):
                 try:
                     try:
                         if isinstance(renderable, CompositorUpdate):
+                            cursor_position = self.screen.size.clamp_offset(
+                                self.cursor_position
+                            )
                             if self._driver.is_inline:
                                 terminal_sequence = Control.move(
                                     *(-self._previous_cursor_position)
                                 ).segment.text
                                 terminal_sequence += renderable.render_segments(console)
                                 terminal_sequence += Control.move(
-                                    *self.cursor_position
+                                    *cursor_position
                                 ).segment.text
                             else:
                                 terminal_sequence = renderable.render_segments(console)
                                 terminal_sequence += Control.move_to(
-                                    *self.cursor_position
+                                    *cursor_position
                                 ).segment.text
-                            self._previous_cursor_position = self.cursor_position
+                            self._previous_cursor_position = cursor_position
                         else:
                             segments = console.render(renderable)
                             terminal_sequence = console._render_buffer(segments)